refactor: extract view config from action config
it should be persisting per topic, and it makes more sense as affecting
the current view.
