+++ src/web/view/actionConfigStore.ts
@@ -2,21 +2,15 @@ import { create } from "zustand";
 import { persist } from "zustand/middleware";
 
 interface ActionConfigStoreState {
-  showImpliedEdges: boolean;
   unrestrictedEditing: boolean;
-  forceNodesIntoLayers: boolean;
   flashlightMode: boolean;
   fillNodesWithColor: boolean;
-  layoutThoroughness: number;
 }
 
 const initialState: ActionConfigStoreState = {
-  showImpliedEdges: false,
   unrestrictedEditing: false,
-  forceNodesIntoLayers: true,
   flashlightMode: false,
   fillNodesWithColor: true,
-  layoutThoroughness: 1,
 };
 
 const useActionConfigStore = create<ActionConfigStoreState>()(
@@ -26,16 +20,8 @@ const useActionConfigStore = create<ActionConfigStoreState>()(
 );
 
 // hooks
-export const useShowImpliedEdges = () => {
-  return useActionConfigStore((state) => state.showImpliedEdges);
-};
-
 export const useUnrestrictedEditing = () => {
   return useActionConfigStore((state) => state.unrestrictedEditing);
-};
-
-export const useForceNodesIntoLayers = () => {
-  return useActionConfigStore((state) => state.forceNodesIntoLayers);
 };
 
 export const useFlashlightMode = () => {
@@ -46,21 +32,9 @@ export const useFillNodesWithColor = () => {
   return useActionConfigStore((state) => state.fillNodesWithColor);
 };
 
-export const useLayoutThoroughness = () => {
-  return useActionConfigStore((state) => state.layoutThoroughness);
-};
-
 // actions
-export const toggleShowImpliedEdges = (show: boolean) => {
-  useActionConfigStore.setState({ showImpliedEdges: show });
-};
-
 export const toggleUnrestrictedEditing = (unrestricted: boolean) => {
   useActionConfigStore.setState({ unrestrictedEditing: unrestricted });
-};
-
-export const toggleForceNodesIntoLayers = (force: boolean) => {
-  useActionConfigStore.setState({ forceNodesIntoLayers: force });
 };
 
 export const toggleFlashlightMode = (flashlight: boolean) => {
@@ -69,10 +43,6 @@ export const toggleFlashlightMode = (flashlight: boolean) => {
 
 export const toggleFillNodesWithColor = (fill: boolean) => {
   useActionConfigStore.setState({ fillNodesWithColor: fill });
-};
-
-export const setLayoutThoroughness = (thoroughness: number) => {
-  useActionConfigStore.setState({ layoutThoroughness: thoroughness });
 };
 
 // utils
